+++ plugins/examples/textured_shape.cxx
@@ -97,7 +97,7 @@ public:
 	}
 	bool init(cgv::render::context& ctx)
 	{
-		size_t di = mat.add_image_file("res://alhambra.png");
+		int di = mat.add_image_file("res://alhambra.png");
 		mat.set_diffuse_index(di);
 		if (mat.ensure_textures(ctx))
 			t_ptr = mat.get_texture(di);
